+++ protocol/client/video.c
@@ -52,7 +52,7 @@ extern volatile bool update_mbps;
 extern volatile int output_width;
 extern volatile int output_height;
 extern volatile CodecType output_codec_type;
-extern volatile float latency;
+extern volatile double latency;
 
 #if CAN_UPDATE_WINDOW_TITLEBAR_COLOR
 extern volatile FractalRGBColor* native_window_color;
